// web/lib/api/generated/models/stocktakeResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * GroceryManager.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StocktakeEntryResponse } from './stocktakeEntryResponse';
import type { StocktakeStatus } from './stocktakeStatus';

export interface StocktakeResponse {
  id: string;
  /** @nullable */
  shoppingPresetId: string | null;
  status: StocktakeStatus;
  startedAtUtc: string;
  /** @nullable */
  completedAtUtc: string | null;
  entries: StocktakeEntryResponse[];
  version: string;
}